add service notices changes to ui
