+++ web/src/app/alerts/AlertsList.tsx
@@ -18,7 +18,7 @@ import CreateFAB from '../lists/CreateFAB'
 import CreateAlertDialog from './CreateAlertDialog/CreateAlertDialog'
 import { useURLParam } from '../actions'
 import { ControlledPaginatedListAction } from '../lists/ControlledPaginatedList'
-import ServiceMaintenanceNotice from '../services/ServiceMaintenanceNotice'
+import ServiceNotices from '../services/ServiceNotices'
 import { Time } from '../util/Time'
 
 interface AlertsListProps {
@@ -264,7 +264,7 @@ export default function AlertsList(props: AlertsListProps): JSX.Element {
   return (
     <React.Fragment>
       <Grid container direction='column' spacing={2}>
-        <ServiceMaintenanceNotice serviceID={props.serviceID} />
+        <ServiceNotices serviceID={props.serviceID} />
         <Grid item>
           <QueryList
             query={alertsListQuery}
